Modified SMC1 transmit behaviour
If an exception occurred, and since I raise the IPL to level 7 in my exception handlers, this meant no more SMC interrupts coule be recognised.

Switched to using the SWT to periodically flush the transmit buffer since it generates interrupts at IRQ7.

Also had issues using pointers for the head and tail, so switched to integer indexes instead.